+++ lib/appium_lib_core/common/base/bridge/w3c.rb
@@ -62,6 +62,12 @@ module Appium
             ::Selenium::WebDriver::Remote::W3C::Capabilities.json_create execute(:get_capabilities)
           end
 
+          # Port from MJSONWP
+          def send_keys_to_active_element(key)
+            text = ::Selenium::WebDriver::Keys.encode(key).join('')
+            execute :send_keys_to_active_element, {}, { value: text.split(//) }
+          end
+
           # For Appium
           # override
           def page_source