+++ lib/appium_lib_core/ios/uiautomation/bridge.rb
@@ -4,7 +4,6 @@ module Appium
       module Uiautomation
         module Bridge
           def self.for(target)
-            Core::Ios::SearchContext.extend
             target.extend Appium::Core::Ios::Device
 
             Core::Ios::Uiautomation.patch_webdriver_element