+++ lib/appium_lib_core/ios/xcuitest/bridge.rb
@@ -4,8 +4,6 @@ module Appium
       module Xcuitest
         module Bridge
           def self.for(target)
-            Core::Ios::SearchContext.extend
-            Core::Ios::Xcuitest::SearchContext.extend
             target.extend Appium::Core::Ios::Device
             target.extend Appium::Core::Ios::Xcuitest::Device
           end