+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Enhancements
 - Update documentation about `start_recording_screen`
+- Port `send_keys/type` for active element
 
 ### Bug fixes