+++ CHANGELOG.md
@@ -3,7 +3,21 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Enhancements
+- Update documentation about `start_recording_screen`
+- Port `send_keys/type` for active element
+- Support `find_element/s :image, partial_image`
+
+### Bug fixes
+
+### Deprecations
+- Deprecate experimental `ImageElement` in favor of `Element`
+    - Anyone does not need to care this
+
+## [1.8.4] - 2018-07-28
+### Enhancements
 - silence warning for pointeractions [#113](https://github.com/appium/ruby_lib_core/pull/113)
+- Use method missing to get attributes like `e.resource_id` instead of `e.attribute 'resource-id'` [#116](https://github.com/appium/ruby_lib_core/pull/116)
+- Set `'~> 3.5', '< 3.14'`
 
 ### Bug fixes